add temp_total_count to Driver
station might not complete all timeline when client first login, to let
client know the actual total post count, station needs to save the value
to temp_total_count and tell client before completed timeline.

TODO: winclient needs to be modified so that it can keep pulling timeline
if it doensn't have as many posts as the returned total count.