win client: check quota only in cloud mode because station does this
Quota rule is changed: only original photo is counted, thumbnails are not.
Win client is free to upload original photo to station which in turns
uploads medium thumbnails to cloud. Station will then upload original
photos only if quota is not full. If quota is full, station waits until
the quota period is over and uploads again.